+++ tests/test_second_checkout_page.py
@@ -1,9 +1,7 @@
 class TestSecondCheckout:
 
     def test_checkout_with_random_data(self, start_page, random_user):
-        start_page.second_checkout(random_user)
-        start_page.verify_checkout_complete()
-
-    def test_total_price_without_add(self, start_page, random_user):
-        start_page.second_checkout(random_user)
-        start_page.verify_total_checkout_without()
+        second_checkout = start_page.navigate_to_second_checkout(random_user)
+        second_checkout.fill_users_data_2(random_user)
+        second_checkout.verify_checkout_complete()
+        second_checkout.verify_total_checkout_without()
